refactor: Lift the live-alert rendering out of the alert
live_alert.py held two things: what an alert looks like, and the lifecycle that
posts, refreshes and closes it. The rendering shares nothing with the lifecycle
but the data it reads, and it is the part somebody changes when they want the
alert to say something different.

live_alert_embeds.py takes the nine functions that turn what Helix returned into
something Discord can post. Six of them become that module's interface and lose
their leading underscore; _title, _display_name, _game and the link escaping stay
private, being used only by the offline embed and each other. Bodies lifted by
AST span, not retyped.

659 -> 488 and 191. Being honest about the number: 488 is still over the
four-hundred line standard, and getting under it would mean cutting the updater
cycle - _refresh, _close, _cycle, _run - which reads as one sequence and is the
module's whole subject. The standard found three real muddles on this branch and
in controller/twitch.py before it; this is the file where it does not fit, and
that is recorded rather than dressed up.

Verified: all three embeds build with the fields they had; a hostile stream title
still cannot escape the link in either online description; the offline
description leaves "](" as literal text, which is right because no bracket
opens before it; the watch button carries its url and the role ping still only
goes to the alerts channel. The webhook harness is unchanged.
